+++ src/components/students/activate-archive-student-modal.js
@@ -1,0 +1,31 @@
+import React from 'react';
+import DeleteIcon from '@mui/icons-material/Delete';
+import GenericModal from '../shared/generic-modal';
+
+export function ArchiveStudentModal() {
+  return (
+    <GenericModal
+      openModal={<DeleteIcon />}
+      modalHeadingTitle="Archive Student"
+      modalMessage="Are you sure you want to archive this student?"
+      actionButtonColor="archive"
+      cancelButtonColor="cancel"
+      actionButtonTitle="Archive"
+      cancelButtonTitle="Cancel"
+    />
+  );
+}
+
+export function ActivateStudentModal() {
+  return (
+    <GenericModal
+      openModal="Accept"
+      modalHeadingTitle="Accept Student"
+      modalMessage="Are you sure you want to accept this student?"
+      actionButtonColor="submit"
+      cancelButtonColor="cancel"
+      actionButtonTitle="Accept"
+      cancelButtonTitle="Cancel"
+    />
+  );
+}
